+++ api/lib/application/organization-learners/index.js
@@ -1,0 +1,73 @@
+const Joi = require('joi').extend(require('@joi/date'));
+
+const securityPreHandlers = require('../security-pre-handlers');
+const organizationLearnerController = require('./organization-learner-controller');
+const identifiersType = require('../../domain/types/identifiers-type');
+
+exports.register = async function (server) {
+  const adminRoutes = [
+    {
+      method: 'DELETE',
+      path: '/api/admin/organization-learners/{id}/association',
+      config: {
+        pre: [
+          {
+            method: (request, h) =>
+              securityPreHandlers.userHasAtLeastOneAccessOf([
+                securityPreHandlers.checkUserHasRoleSuperAdmin,
+                securityPreHandlers.checkUserHasRoleSupport,
+              ])(request, h),
+            assign: 'hasAuthorizationToAccessAdminScope',
+          },
+        ],
+        handler: organizationLearnerController.dissociate,
+        validate: {
+          params: Joi.object({
+            id: identifiersType.organizationLearnerId,
+          }),
+        },
+        notes: [
+          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
+            '- Elle dissocie un utilisateur d’un prescrit',
+        ],
+        tags: ['api', 'admin', 'organization-learners'],
+      },
+    },
+  ];
+
+  server.route([
+    ...adminRoutes,
+    {
+      method: 'DELETE',
+      path: '/api/schooling-registration-user-associations/{id}',
+      config: {
+        pre: [
+          {
+            method: (request, h) =>
+              securityPreHandlers.userHasAtLeastOneAccessOf([
+                securityPreHandlers.checkUserHasRoleSuperAdmin,
+                securityPreHandlers.checkUserHasRoleCertif,
+                securityPreHandlers.checkUserHasRoleSupport,
+                securityPreHandlers.checkUserHasRoleMetier,
+              ])(request, h),
+            assign: 'hasAuthorizationToAccessAdminScope',
+          },
+        ],
+        handler: organizationLearnerController.dissociate,
+        validate: {
+          params: Joi.object({
+            id: identifiersType.schoolingRegistrationId,
+          }),
+        },
+        notes: [
+          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
+            '- Elle dissocie un utilisateur d’une inscription d’élève',
+          "- L'usage de cette route est **dépréciée** en faveur de /api/admin/organization-learners/{id}/association",
+        ],
+        tags: ['api', 'admin', 'schoolingRegistrationUserAssociation'],
+      },
+    },
+  ]);
+};
+
+exports.name = 'organization-learners-api';
