+++ api/lib/routes.js
@@ -39,6 +39,7 @@ module.exports = [
   require('./application/saml'),
   require('./application/organization-learner-dependent-users'),
   require('./application/organization-learner-user-associations'),
+  require('./application/organization-learners'),
   require('./application/scorecards'),
   require('./application/sessions'),
   require('./application/simulateErrors'),
