+++ api/tests/acceptance/application/organization-learner-controller_test.js
@@ -1,0 +1,66 @@
+const {
+  databaseBuilder,
+  expect,
+  insertUserWithRoleSuperAdmin,
+  generateValidRequestAuthorizationHeader,
+} = require('../../test-helper');
+
+const createServer = require('../../../server');
+
+describe('Acceptance | Controller | organization-learner', function () {
+  let server;
+
+  beforeEach(async function () {
+    server = await createServer();
+  });
+
+  describe('DELETE /api/schooling-registration-user-associations/{id}', function () {
+    context('When user has the role SUPER_ADMIN and organization learner can be dissociated', function () {
+      it('should return an 204 status after having successfully dissociated user from organizationLearner', async function () {
+        const organizationId = databaseBuilder.factory.buildOrganization({ isManagingStudents: true }).id;
+        const superAdmin = await insertUserWithRoleSuperAdmin();
+        const userId = databaseBuilder.factory.buildUser().id;
+        const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({ organizationId, userId });
+
+        await databaseBuilder.commit();
+
+        const options = {
+          method: 'DELETE',
+          url: `/api/schooling-registration-user-associations/${organizationLearner.id}`,
+          headers: {
+            authorization: generateValidRequestAuthorizationHeader(superAdmin.id),
+          },
+        };
+
+        const response = await server.inject(options);
+
+        expect(response.statusCode).to.equal(204);
+      });
+    });
+  });
+
+  describe('DELETE /api/admin/organization-learners/{id}/association', function () {
+    context('When user has the role SUPER_ADMIN and organization learner can be dissociated', function () {
+      it('should return an 204 status after having successfully dissociated user from organizationLearner', async function () {
+        const organizationId = databaseBuilder.factory.buildOrganization({ isManagingStudents: true }).id;
+        const superAdmin = await insertUserWithRoleSuperAdmin();
+        const userId = databaseBuilder.factory.buildUser().id;
+        const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({ organizationId, userId });
+
+        await databaseBuilder.commit();
+
+        const options = {
+          method: 'DELETE',
+          url: `/api/admin/organization-learners/${organizationLearner.id}/association`,
+          headers: {
+            authorization: generateValidRequestAuthorizationHeader(superAdmin.id),
+          },
+        };
+
+        const response = await server.inject(options);
+
+        expect(response.statusCode).to.equal(204);
+      });
+    });
+  });
+});
