+++ api/tests/acceptance/application/organization-learner-user-association-controller_test.js
@@ -1,7 +1,6 @@
 const {
   databaseBuilder,
   expect,
-  insertUserWithRoleSuperAdmin,
   generateValidRequestAuthorizationHeader,
   generateIdTokenForExternalUser,
   knex,
@@ -1316,56 +1315,6 @@ describe('Acceptance | Controller | organization-learner-user-associations', fun
           // then
           expect(response.statusCode).to.equal(422);
         });
-      });
-    });
-  });
-
-  describe('DELETE /api/schooling-registration-user-associations', function () {
-    context('When user has the role SUPER_ADMIN and organization learner can be dissociated', function () {
-      it('should return an 204 status after having successfully dissociated user from organizationLearner', async function () {
-        const organizationId = databaseBuilder.factory.buildOrganization({ isManagingStudents: true }).id;
-        const superAdmin = await insertUserWithRoleSuperAdmin();
-        const userId = databaseBuilder.factory.buildUser().id;
-        const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({ organizationId, userId });
-
-        await databaseBuilder.commit();
-
-        const options = {
-          method: 'DELETE',
-          url: `/api/schooling-registration-user-associations/${organizationLearner.id}`,
-          headers: {
-            authorization: generateValidRequestAuthorizationHeader(superAdmin.id),
-          },
-        };
-
-        const response = await server.inject(options);
-
-        expect(response.statusCode).to.equal(204);
-      });
-    });
-  });
-
-  describe('DELETE /api/admin/schooling-registration-user-associations', function () {
-    context('When user has the role SUPER_ADMIN and organization learner can be dissociated', function () {
-      it('should return an 204 status after having successfully dissociated user from organizationLearner', async function () {
-        const organizationId = databaseBuilder.factory.buildOrganization({ isManagingStudents: true }).id;
-        const superAdmin = await insertUserWithRoleSuperAdmin();
-        const userId = databaseBuilder.factory.buildUser().id;
-        const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({ organizationId, userId });
-
-        await databaseBuilder.commit();
-
-        const options = {
-          method: 'DELETE',
-          url: `/api/admin/schooling-registration-user-associations/${organizationLearner.id}`,
-          headers: {
-            authorization: generateValidRequestAuthorizationHeader(superAdmin.id),
-          },
-        };
-
-        const response = await server.inject(options);
-
-        expect(response.statusCode).to.equal(204);
       });
     });
   });
