+++ api/tests/unit/application/organization-learner-user-associations/index_test.js
@@ -141,153 +141,4 @@ describe('Unit | Application | Router | organization-learner-user-associations-r
       });
     });
   });
-
-  describe('DELETE /api/schooling-registration-user-associations/{id}', function () {
-    const method = 'DELETE';
-
-    it('should return a HTTP status code 204', async function () {
-      // given
-      sinon.stub(securityPreHandlers, 'userHasAtLeastOneAccessOf').returns(() => true);
-      sinon
-        .stub(organizationLearnerUserAssociationController, 'dissociate')
-        .callsFake((request, h) => h.response('ok').code(204));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const url = '/api/schooling-registration-user-associations/1';
-
-      // when
-      const response = await httpTestServer.request(method, url, null);
-
-      // then
-      expect(response.statusCode).to.equal(204);
-    });
-
-    it('should return a HTTP status code 403 when user does not have rights to Pix Admin', async function () {
-      // given
-      sinon
-        .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
-        .returns((request, h) => h.response().code(403).takeover());
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const url = '/api/schooling-registration-user-associations/1';
-
-      // when
-      const response = await httpTestServer.request(method, url, null);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-    });
-
-    it('should return a HTTP status code 400 if id parameter is not a number', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const url = '/api/schooling-registration-user-associations/ABC';
-
-      // when
-      const response = await httpTestServer.request(method, url, null);
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-  });
-
-  context('Routes /admin', function () {
-    describe('DELETE /api/admin/schooling-registration-user-associations/{id}', function () {
-      it('should return a HTTP status code 204 when user role is "SUPER_ADMIN"', async function () {
-        // given
-        sinon.stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin').callsFake((request, h) => h.response(true));
-        sinon
-          .stub(securityPreHandlers, 'checkUserHasRoleSupport')
-          .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-        sinon
-          .stub(organizationLearnerUserAssociationController, 'dissociate')
-          .callsFake((request, h) => h.response('ok').code(204));
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request(
-          'DELETE',
-          '/api/admin/schooling-registration-user-associations/1'
-        );
-
-        // then
-        sinon.assert.calledOnce(securityPreHandlers.checkUserHasRoleSuperAdmin);
-        sinon.assert.calledOnce(securityPreHandlers.checkUserHasRoleSupport);
-        sinon.assert.calledOnce(organizationLearnerUserAssociationController.dissociate);
-        expect(response.statusCode).to.equal(204);
-      });
-
-      it('should return a HTTP status code 204 when user role is "SUPPORT"', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin')
-          .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-        sinon.stub(securityPreHandlers, 'checkUserHasRoleSupport').callsFake((request, h) => h.response(true));
-        sinon
-          .stub(organizationLearnerUserAssociationController, 'dissociate')
-          .callsFake((request, h) => h.response('ok').code(204));
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request(
-          'DELETE',
-          '/api/admin/schooling-registration-user-associations/1'
-        );
-
-        // then
-        sinon.assert.calledOnce(securityPreHandlers.checkUserHasRoleSuperAdmin);
-        sinon.assert.calledOnce(securityPreHandlers.checkUserHasRoleSupport);
-        sinon.assert.calledOnce(organizationLearnerUserAssociationController.dissociate);
-        expect(response.statusCode).to.equal(204);
-      });
-
-      it('should return a HTTP status code 403 when user does not have access (CERTIF | METIER)', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin')
-          .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-        sinon
-          .stub(securityPreHandlers, 'checkUserHasRoleSupport')
-          .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-        sinon
-          .stub(organizationLearnerUserAssociationController, 'dissociate')
-          .callsFake((request, h) => h.response('ok').code(204));
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request(
-          'DELETE',
-          '/api/admin/schooling-registration-user-associations/1'
-        );
-
-        // then
-        sinon.assert.calledOnce(securityPreHandlers.checkUserHasRoleSuperAdmin);
-        sinon.assert.calledOnce(securityPreHandlers.checkUserHasRoleSupport);
-        sinon.assert.notCalled(organizationLearnerUserAssociationController.dissociate);
-        expect(response.statusCode).to.equal(403);
-      });
-
-      it('should return a HTTP status code 400 if id parameter is not a number', async function () {
-        // given
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request(
-          'DELETE',
-          '/api/admin/schooling-registration-user-associations/ABC'
-        );
-
-        // then
-        expect(response.statusCode).to.equal(400);
-      });
-    });
-  });
 });
